Throw on invalid input type; null if no name
Passing a non-function into something that gives you the name of a
function is clearly a user error and should be treated as such.

Returning null for functions that don't have a name makes more sense
than the previous mapping to "anonymous":  You can have a function with
a name "anonymous", which that would be confused with.

It is slightly confusing that the `new Function` constructor returns a
function that has a name, and that name is "anonymous", but eh, it's
JavaScript, what do you expect.